# tia/dua-tia-configuration-manager/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/tia/api/v1/repository-search-bindings": {
            "get": {
                "tags": [
                    "repository-search-binding-controller"
                ],
                "operationId": "getRepositorySearchBindings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RepositorySearchBinding"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "repository-search-binding-controller"
                ],
                "operationId": "saveRepositorySearchBinding",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RepositorySearchBinding"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RepositorySearchBinding"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/tia/api/v1/repository-search-bindings/{id}": {
            "get": {
                "tags": [
                    "repository-search-binding-controller"
                ],
                "operationId": "getRepositorySearchBinding",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RepositorySearchBinding"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "repository-search-binding-controller"
                ],
                "operationId": "deleteRepositorySearchBinding",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/repository-search-bindings/{id}/acronym": {
            "get": {
                "tags": [
                    "repository-search-binding-controller"
                ],
                "operationId": "getRepositorySearchBindingAcronym",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/repository-search-bindings/by-doc-repo/{docRepoId}": {
            "get": {
                "tags": [
                    "repository-search-binding-controller"
                ],
                "operationId": "getRepositorySearchBindingByDocRepo",
                "parameters": [
                    {
                        "name": "docRepoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RepositorySearchBinding"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "RepositorySearchBinding": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "initiativeId": {
                        "type": "string"
                    },
                    "documentTypeId": {
                        "type": "string"
                    },
                    "keywordDictionaryId": {
                        "type": "string"
                    },
                    "documentRepositoryIds": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        }
    }
}